Desktop/CursorGoat/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0
openai==1.51.0 